pyproject.toml
[project]
name = "magictales-background-audio"
version = "0.1.0"
description = "Generate musical loops from book text using Google Vertex AI Lyria"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "typer>=0.12",
    "httpx>=0.27",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "anthropic>=0.40",
    "google-auth>=2.0",
    "google-genai>=1.0",
    "rich>=13.0",
    "pypdf>=4.0",
    "ebooklib>=0.18",
    "beautifulsoup4>=4.12",
    "python-docx>=1.1",
]

[project.optional-dependencies]
gemini = [
    "google-cloud-aiplatform>=1.40",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.5",
    "respx>=0.22",
]

[project.scripts]
magictales = "background_audio.cli:app"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["background_audio*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100
